Increased calibration dot size to loosen saccade threshold
It turns out calibration dot size was being fed to the eyelink_child.py
script as the threshold for reporting a blink or saccade. The previous
dot size (threshold) of 0.5 degrees was very strict.

The reason this did not cause any problems with the Ghis script is that
the eye tracker settings were improperly set such that the tracker
thought the screen was much smaller than it actually was. This meant
that fewer saccades were reported passed to through the
eyelink_child.py script to be evaluated against the 0.5 threshold.

Since, the screen options have been changed, meaning now the dot size
(threshold) need to be adapted so that fewer saccades are detected once
they are passed through the script.